+++ public/js/stock.js
@@ -5,10 +5,3 @@ import realtime from './modules/realtime';
 darkUnicaTheme();
 loadHome();
 realtime();
-
-/*
-script(src="js/loadHome.js")
-script(src="/js/realtime.js")
-script(src="js/stocksTheme.js")
-script(src="js/createChart.js")
-*/
